deleted raymond's file, added new structure for ADM.py
