+++ DetectionScripts/InitMacros.py
@@ -1,0 +1,4 @@
+
+
+def ex():
+    print("hello")
